bug mili->sec 5min. de timeout tirei rnddelay do acq2

+++ src/main/scala/util/Lock.scala
@@ -35,7 +35,7 @@ trait Lock {
     Thread.sleep((rnd.nextDouble() * seconds * 1000d).toInt)
   }
 
-  def checkExistsForNFS(f: File, delay: Int = 0.04) = {
+  def checkExistsForNFS(f: File, delay: Double = 0.04) = {
     rndDelay(delay)
     try {
       val buffer = new Array[Byte](4)
